requirements.txt
python-dotenv==0.18.0
fitter==1.4.0
networkx==2.6.3
fastapi==0.70.0
uvicorn==0.15.0
boto3==1.20.10
botocore==1.23.10
dash==2.0.0
dash-bootstrap-components==1.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.2
great-expectations==0.13.38
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.3
plotly==5.3.1
scipy==1.7.3
uvicorn==0.15.0
starlette==0.16.0

